Implement the commit offset and kafka config holder
