+++ config.example.wikimedia.yaml
@@ -30,8 +30,16 @@ spec: &spec
       x-modules:
         - path: sys/kafka.js
           options:
-            uri: 127.0.0.1:2181
+            metadata_broker_list: 127.0.0.1:9092
             dc_name: test_dc
+            produce_dc: test_dc
+            consumer:
+              # These options should not be copied to puppet config.
+              # We're using this config for testing, so need to configure
+              # for minimal latency
+              fetch.wait.max.ms: "1"
+              fetch.min.bytes: "1"
+              queue.buffering.max.ms: "1"
             concurrency: 250
             templates:
