+++ config.example.yaml
@@ -42,7 +42,7 @@ services:
             x-modules:
               - path: sys/kafka.js
                 options:
-                  uri: 127.0.0.1:2181
+                  metadata_broker_list: 127.0.0.1:9092
                   dc_name: default
                   concurrency: 10
                   templates: {}
